ObjexVision/requirements.txt
numpy
pandas
Pillow
streamlit
tensorflow
requests
streamlit-lottie
matplotlib>=3.0.0
